Update readthedocs config re: new API (#27)
* Change dev env to Python 3.11

Drop support for Python 3.10.

* Pin sphinx-rtd-theme=1.3

We're relying on the sphinx-rtd-theme version pin to ensure compatible versions
of the rest of the Sphinx tool chain.

* pre-commit autoupdate

* Rename readthedocs.yml to .readthedocs.yaml

re: change in readthedocs config API

* Change to use new readthedocs build spec API

* Change cookiecutter context paths to strings

As of Python 3.11, the standard library json module refuses to serialize
pathlib.Path objects. So, all paths in cookiecutter context have to be wrapped
in os.fspath().