Docs maintenance (#48)
* Update ReadTheDocs build config

Upgraded the OS to Ubuntu 24.04 and updated the Python toolchain to
mambaforge-23.11 in the .readthedocs.yaml file. This ensures
compatibility with the latest environment settings available on
readthedocs.

* Update Sphinx & optional dependencies version pin

Updated Sphinx and optional dependencies packages pinned versions in
`environment-dev.yaml`, `environment-test.yaml`, and `environment-rtd.yaml` in
preparation for Python 3.13 environment.

* Update redirected docs links from linkcheck

Updated the URLs for Conda references in `atlantis_cmd.rst` and
`pkg_development.rst` to point to the correct latest documentation. This ensures
users are directed to the accurate and updated official Conda resources.

* Update docs example output from Sphinx and pytest

Updated Sphinx build and linkcheck commands example outputs. Also revised the
pytest output to reflect the newer plugin versions and test suite logging
changes.