Modernize packaging (#49)
* Move project metadata & entry points to pyproject.toml

Transitioned from setup.py and setup.cfg to pyproject.toml for managing project
configuration, aligning with modern standards in Python packaging. This change
simplifies dependency management and enhances compatibility with build tools.
Adjusted command entry points accordingly in the main module documentation.

* Set static version to 22.1.dev0 in pyproject.toml

Updated the version attribute in pyproject.toml to a static value, "22.1.dev0",
from the previously dynamic setting. This fixes an issue in the GitHub
Actions CI workflows. That issue is expected to be transient while this PR is
in progress.

* Change to Hatch for package management

Replaced `setuptools` with `hatchling` in `pyproject.toml` for package
management and build system. Updated development environment configuration and
documentation to reflect Hatch's integration, along with adding a Hatch badge to
the README and development docs.

* Move version identifier to __about__.py

Moved version information from `atlantis_cmd/__init__.py` to a new
`atlantis_cmd/__about__.py` file. The `pyproject.toml` is updated to make
versioning dynamic by referencing the new path for versioning. This change
centralizes version management and keeps the init file cleaner.

* Update version retrieval in docs configuration

Replaced configuration parsing with `importlib.metadata.version()`` for version
retrieval. This change simplifies the code and aligns the versioning method with
the latest standards in Python. Additionally, it removes the dependency on the
atlantis_cmd module for collecting the version information.

* Update version identifier to exclude it from coverage

Added a pragma directive to the version string to exclude it from test coverage
metrics. This ensures that changes to the version number do not affect code
coverage reports.

* Add "Release Process" section to documentation

This commit introduces a comprehensive section outlining the release process in
the package development documentation. It includes step-by-step instructions on
preparing and publishing new releases, managing milestones, and updating version
numbers. This addition aims to standardize the release workflow and ensure
consistency in future releases.

* Change badges layout in README & dev docs to table

Reorganized the badges in README.rst and package development documentation files
into structured sections for better readability and maintainability. Added new
pre-commit badge and updated instructions for installing pre-commit hooks in the
development environment setup.

* Update pkgs & versions used in recent dev env