Change to repo-specific pytest-with-coverage workflow
Replace the reusable workflow with a repo-specific one. This is necessary to
work around an issue that results in coverage failing with:
    INTERNALERROR> coverage.exceptions.DataError: Can't combine statement
    coverage data with branch data
The root cause seems to be that something in the repo (maybe cookiecutter)
causes the tests to run in parallel on 1 core and coverage gets confused.
Using `--cov-config=pyproject.toml` resolves the issue.